+++ AVSCommon/AVS/test/Attachment/AttachmentReaderTest.cpp
@@ -122,7 +122,7 @@ void AttachmentReaderTest::testMultipleReads(bool closeWriterBeforeReading) {
     std::vector<uint8_t> result(TEST_SDS_PARTIAL_READ_AMOUNT_IN_BYTES);
     auto readStatus = InProcessAttachmentReader::ReadStatus::OK;
 
-    int totalBytesRead = 0;
+    size_t totalBytesRead = 0;
     bool done = false;
     int iterations = 0;
     int iterationsMax = 10;
@@ -145,7 +145,7 @@ void AttachmentReaderTest::testMultipleReads(bool closeWriterBeforeReading) {
     // Not only was all the data read, but the reader remained open.
     ASSERT_NE(iterations, iterationsMax);
     ASSERT_EQ(readStatus, terminalStatus);
-    ASSERT_EQ(totalBytesRead, static_cast<ssize_t>(m_testPattern.size()));
+    ASSERT_EQ(totalBytesRead, m_testPattern.size());
 }
 
 void AttachmentReaderTest::readAndVerifyResult(
